refactor: :recycle: Ajustei nomenclatura
