+++ src/shared/types/planta.d.ts
@@ -1,0 +1,5 @@
+export interface PlantaPreview {
+    id: number;
+    name: string;
+    foto?: string;
+}
